Refactoring for displaying user's info

+++ Rotate.ino
@@ -1,6 +1,6 @@
 #include <Servo.h>
 
-const unsigned int MAX_ANGLE = 60;
+const byte MAX_ANGLE = 60;
 
 // Motor
 Servo servo;  // create servo object to control a servo
@@ -17,7 +17,7 @@ void initServo(){
 void handleRotate() { 
   /*int minutes = duration / 60;
   int hours = duration / (60 * 60);*/
-  const int certianMinute = 7;
+  const byte certianMinute = 7;
   if (time.Hours % 2 == 0 && time.minutes == certianMinute) {
     rotateLeft();  
   } else if (time.Hours % 2 == 1 && time.minutes == certianMinute) {
@@ -26,7 +26,7 @@ void handleRotate() {
 }
 
 void rotateLeft() {  
-  int startAngle = constrain(servo.read(), 0, MAX_ANGLE);
+  byte startAngle = constrain(servo.read(), 0, MAX_ANGLE);
   boolean isStartLimitAngle = startAngle < 10;
   if (isStartLimitAngle) {
     Serial.print(time.gettime("d-m-Y, H:i:s,"));
@@ -34,7 +34,7 @@ void rotateLeft() {
     Serial.println(startAngle);
 
     servo.attach(SERVO_SLOT); 
-    for (int angle = startAngle; angle <= MAX_ANGLE; angle += 1) { // goes from 0 degrees to MAX_ANGLE degrees
+    for (byte angle = startAngle; angle <= MAX_ANGLE; angle += 1) { // goes from 0 degrees to MAX_ANGLE degrees
       servo.write(angle);              // tell servo to go to position in variable 'pos'
       delay(100);                      // waits 100ms for the servo to reach the position
     }
@@ -45,7 +45,7 @@ void rotateLeft() {
 
 
 void rotateRight() {
-  int startAngle = constrain(servo.read(), 0, MAX_ANGLE);
+  byte startAngle = constrain(servo.read(), 0, MAX_ANGLE);
   boolean isStartLimitAngle = abs(startAngle - MAX_ANGLE) < 10;
   if (isStartLimitAngle) {
     Serial.print(time.gettime("d-m-Y, H:i:s,"));
@@ -53,7 +53,7 @@ void rotateRight() {
     Serial.println(startAngle);
     
     servo.attach(SERVO_SLOT);
-    for (int angle = startAngle; angle >= 0; angle -= 1) { // goes from MAX_ANGLE degrees to 0 degrees
+    for (byte angle = startAngle; angle >= 0; angle -= 1) { // goes from MAX_ANGLE degrees to 0 degrees
       servo.write(angle);              // tell servo to go to position in variable 'pos'
       delay(100);                      // waits 100ms for the servo to reach the position
     }
